Add files via upload
Changed how we got the values so would not have to & 0xff with every read from memory

+++ Memory/MemRegisters.java
@@ -21,7 +21,7 @@ public class MemRegisters {
         return value&0xff;
     }
 
-    public void setByte(byte value) {
+    public void setByte(int value) {
         this.value = value;
     }
